Commit13

+++ src/FixPaymentWorker.java
@@ -1,8 +1,3 @@
-
-import java.text.NumberFormat;
-import java.util.Locale;
-
-
 public class FixPaymentWorker extends Worker{
     private int monthPayment;
 
@@ -12,7 +7,7 @@ public class FixPaymentWorker extends Worker{
     }
 
     @Override
-    double getAveragePayment() {
+    int getAveragePayment() {
         return monthPayment;
     }
